# com.trekglobal.idempiere.rest.api/openapi/idempiere-rest.yml
openapi: 3.0.0
info:
  title: iDempiere REST endpoints
  description: >-
    iDempiere REST endpoints.
  version: 1.0.0
servers:
  - url: '{base_url}'
    variables:
      base_url:
        enum:
          - 'http://localhost:8080/api/v1'
          - 'https://localhost:8443/api/v1'
          - 'https://test.idempiere.org/api/v1'
        default: 'http://localhost:8080/api/v1'
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes  
  parameters:
...
    filter:
      in: query
      name: $filter
      required: false
      schema:
        type: string
      description: query filter
    orderby:
      in: query
      name: $orderby
      required: false
      schema:
        type: string
      description: ordering of query result
...
  responses:
...
    ServerError:
      description: Internal Server Error. Unhandled exceptions.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internal_server_error:
              value:
                title: "Internal Server Error"
                status: 500
                detail: "Internal server error. Unhandled exceptions."
...
    ForbiddenGetRecordRequest:
      description: Forbidden. Access to this specific record instance is restricted.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_get_record:
              value:
                title: "Access denied"
                status: 403
                detail: "Access denied for table: Products"
...
  schemas:
...
    ErrorResponse:
      description: Error response from server
      type: object
      properties:
        title:
          type: string
          description: error summary
        status:
          type: integer
          description: http status code
        detail:
          type: string
          description: error details
    Models:
      type: object
      properties:
        models:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                description: The unique numeric identifier for the model.
              uid:
                type: string
                format: uuid # Recommended to specify UUID format
                description: The unique UUID identifier for the model.
              model-name:
                type: string
                description: The machine-readable name of the model (e.g., database table name).
              name:
                type: string
                description: The human-readable name of the model.
              description:
                type: string
                description: A description of the model.
          description: An array of model definitions.
      description: A collection of models (AD_Table).
...
security:
    - bearerAuth: [] # use the same name as above

# The x-* below is for reference only    
...
paths:
...
  /models:
    get:
      tags:
        - Models
      summary: Get list of model
      parameters:
        - $ref: '#/components/parameters/filter'
      responses:
        '200':
          description: OK. Returns the list of models.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models'
        '400':
          description: Bad Request. Syntax error in the $filter query parameter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_filter:
                  value:
                    title: 'GET Error'
                    status: 400
                    detail: 'Unsupported operator: xyz'
        '500':
          $ref: '#/components/responses/ServerError'
  /models/{tableName}/yaml:
    get:
      tags:
        - Models
      summary: Get model yaml schema by table name
      parameters:
        - name: tableName
          in: path
          schema:
            type: string
          required: true
          description: table name
          example: 'ad_user'
      responses:
        '200':
          description: OK. Returns the yaml schema of the table.
          content:
            application/yaml:
              schema:
                type: string
                description: yaml schema of table
        '403':
          $ref: '#/components/responses/ForbiddenGetRecordRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /models/{tableName}:
    parameters:
      - name: tableName
        in: path
        schema:
          type: string
        description: table name
        required: true
        examples: 
          C_Tax:
            value: 'c_tax'
          C_BPartner:
            value: 'c_bpartner'
          C_UOM:
            value: 'c_uom'
          C_Order:
            value: 'c_order'
      - $ref: '#/components/parameters/locale'
    get:
      tags:
        - Models
      summary: Get records for table.
      description: Get records for table with optional filter, orderBy and pageNo parameter. Default page size is 100.
      parameters:
        - name: $expand
          in: query
          required: false
          schema:
            type: string
          examples:
            noExpand:
              value: ''
            C_OrderTax_C_OrderLine:
              value: C_OrderTax,C_OrderLine($select=Line,Linenetamt ; $top=1 ; $skip=1 ; $filter=IsActive eq true ; $orderby=Line)            
        - name: $filter
          in: query
          required: false
          schema:
            type: string
          examples:
            noFilter:
              value: ''
            inOperator_C_Tax:
              value: Name in ('106','PST','Exempt') AND C_Tax_ID in (106,107)
            equalOperator_C_Tax:
              value: Name eq 'Exempt'
            C_Tax_C_TaxCategory_ID_107:
              value: C_TaxCategory_ID eq 107
        - $ref: '#/components/parameters/orderby'
        - name: $select
          in: query
          required: false
          schema:
            type: string
          examples:
            noSelect:
              value: ''
            name:
              value: Name
            C_Order_DocumentNo_Description:
              value: DocumentNo,Description
        - $ref: '#/components/parameters/top'
        - $ref: '#/components/parameters/skip'
        - name: $valrule
          in: query
          required: false
          schema:
            type: integer
          examples:
            noValRule:
              value: 
            valRule_C_UOM:
              value: 210
        - name: $context
          in: query
          required: false
          schema:
            type: string
          examples:
            noContext: 
              value:
            valRule_C_UOM_210:
              value: 'M_Product_ID:124'
        - $ref: '#/components/parameters/showsql'
        - $ref: '#/components/parameters/label'
        - $ref: '#/components/parameters/showlabel'
      responses:
        '200':
          description: OK. Returns the requested list of records and pagination metadata.
          content:
            application/json:
              schema:
                type: object
                properties:
                  page-count:
                    type: integer
                    description: number of pages
                    example: 1
                  records-size:
                    type: integer
                    description: number of records per page
                    example: 10
                  skip-records:
                    type: integer
                    description: number of skipped records
                    example: 0
                  row-count:
                    type: integer
                    description: total number of records
                    example: 10
                  array-count:
                    type: integer
                    description: number of arrays
                    example: 1
                  records:
                    type: array
                    description: Records
                    items:
                      type: object
        '400':
          $ref: '#/components/responses/BadQueryRecordsRequest'
        '403':
          $ref: '#/components/responses/ForbiddenQueryRecordsRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '501':
          $ref: '#/components/responses/NotImplementedQueryMethod'
    post:
      tags:
        - Models
      summary: Create new record for table
      requestBody:
        content:
          application/json:
            examples:
              C_Tax:
                value:
                  {
                    "AD_Org_ID": {
                        "id": 0,
                        "tableName": "AD_Org"
                    },
                    "isActive": true,
                    "name": "MyGST",
                    "description": "MyCanadian Federal Sales Tax",
                    "Parent_Tax_ID": {
                        "identifier": "GST/PST",
                        "tableName": "C_Tax"
                    },
                    "C_Country_ID": {
                        "identifier": "Canada",
                        "tableName": "C_Country"
                    },
                    "To_Country_ID": {
                        "identifier": "Canada",
                        "tableName": "C_Country"
                    },
                    "C_TaxCategory_ID": {
                        "identifier": "Standard",
                        "tableName": "C_TaxCategory"
                    },
                    "isDocumentLevel": false,
                    "validFrom": "2001-01-01",
                    "isSummary": false,
                    "rate": "7.0",
                    "requiresTaxCertificate": false,
                    "taxIndicator": "GST",
                    "isDefault": false,
                    "isTaxExempt": false,
                    "sOPOType": {
                        "id": "B"
                    },
                    "isSalesTax": false,
                    "tableName": "C_Tax"
                }
              C_BPartner_User_ContactInterest:
                value:
                  {
                      "name":"Test creating BP with User and Interest Area from REST",
                       "AD_User": [
                           {
                               "Name":"Test User with Interest area",
                               "R_ContactInterest": [
                                   {
                                       "R_InterestArea_ID":101
                                   }
                               ]
                           }
                       ]
                  }
      responses:
        '201':
          description: Created. Returns the created record.
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadCreateRecordRequest'
        '403':
          $ref: '#/components/responses/ForbiddenCreateRecordRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /models/{tableName}/{id}:
    parameters:
      - name: tableName
        in: path
        schema:
          type: string
        description: table name
        required: true
        examples:
          AD_User:
            value: 'ad_user'
          C_Invoice:
            value: 'c_invoice'
          C_BPartner:
            value: 'c_bpartner'
          C_Tax:
            value: 'c_tax'
          C_Tax_Acct:
            value: 'c_tax_acct'
          M_Product:
            value: 'm_product'
          C_BPartner_Location:
            value: 'c_bpartner_location'
          C_Cash:
            value: 'C_Cash'
      - name: id
        in: path
        schema:
          oneOf:
            - type: integer
            - type: string
        description: integer record id or string record uuid
        required: true
        examples:
          AD_User_100:
            value: 100
          AD_User_102:
            value: 102
          C_Invoice_104:
            value: 104
          C_BPartner_119:
            value: 119
          C_BPartner_Location_117:
            value: 117
          C_Tax_106:
            value: 106
          C_Tax_1000000:
            value: 1000000
          M_Product_122:
            value: 122
          C_Tax_Acct_UUID:
            value: 7df64952-88e0-4ebf-9f56-9cb899b555f8
          C_Cash_101:
            value: 101
      - $ref: '#/components/parameters/locale'
    get:
      tags:
        - Models
      summary: Get record by id.
      parameters:                
        - name: $expand
          in: query
          required: false
          schema:
            type: string
          examples:
            noExpand:
              value: ''
            AD_User_100_C_Order.SalesRep:
              value: C_Order.salesrep_id($select=documentno)
            C_Invoice_104_FactAcct_InvoiceLine:
              value: fact_acct.record_id($select=fact_acct_id),c_invoiceline
            C_BPartner_119_AD_User_R_ContactInterest:
              value: AD_User($select=Name ; $expand=R_ContactInterest($select=IsActive ; $filter=IsActive eq true))
            C_Tax_106_C_Tax_Acct:
              value: c_tax_acct
            M_Product_122_ProductCategoryID_ProductPrice:
              value: m_product_category_id($select=Name,IsDefault),M_ProductPrice
        - name: $select
          in: query
          required: false
          schema:
            type: string
          examples:
            noSelect:
              value: ''
            name:
              value: Name
        - $ref: '#/components/parameters/showsql'
        - $ref: '#/components/parameters/showlabel'
      responses:
        '200':
          description: OK. Returns the record data.
          content:
            application/json: {}
        '403':
          $ref: '#/components/responses/ForbiddenGetRecordRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - Models
      summary: update record by id
      requestBody:
        content:
          application/json:
            examples:
              C_Tax_1000000:
                value:
                  {
                      "id": 1000000,
                      "name": "MyGST",
                      "description": "My Sales Tax",
                      "validFrom": "2019-01-01",
                      "rate": "6.0",
                      "isSalesTax": true,
                      "taxIndicator": "GST",
                      "tableName": "C_Tax"
                  }
              C_BPartner_Location_117_Location:
                value:
                  {
                      "Name": "Monroe",
                      "C_Location_ID": {
                          "id": 125,  # if id = 0 a new C_Location is created an linked to the record
                          "Address1": "address 1",
                          "Address2": "address 2",
                          "Address3": "address 3",
                          "Address4": "address 4",
                          "Address5": "address 5",
                          "Postal": "postal",
                          "Postal_Add": "add",
                          "C_Country_ID": 100,
                          "C_Region_ID": 102,
                          "RegionName": "RN",
                          "City": "Near Village",
                          "Comments": "comments",
                          "IsActive": true
                      }
                  }
              AD_User_102_Image:
                value:
                  {
                     "AD_Image_ID":{
                        "file_name":"idempiereLogo.jpg",
                        "url":"org/idempiere/idempiereLogo.jpg",
                        "data":"/9j/4AAQSkZJRgABAQAAAQABAAD/2wCEAAkGBxITEhUSExEVFhMWGCAXGRcWGBgZFxodHBgeGBobHx8fHiggHR8mHh0YJTEjJikrLi4uICAzODMuQygtLisBCgoKDg0OGxAQGy0lHyUtLzcrNS0tNzcyNjcrLS0tNTUtNzctKy0vMjUtLTgvMjYtLS01MS4tMDc3LS0tKystK//AABEIAIUAjgMBIgACEQEDEQH/xAAcAAEAAwADAQEAAAAAAAAAAAAABQYHAQMEAgj/xAA/EAACAQMCAwUFBQUGBwAAAAABAgMABBEFEgYhMRNBUWFxByIygZEUcqGx0TVCUsHhJDM2g7PxI2Jjc4Ky8P/EABoBAQADAQEBAAAAAAAAAAAAAAACAwQGAQX/xAAjEQEAAgICAgICAwAAAAAAAAAAAQIDBBEhE0ESMWGRBSJR/9oADAMBAAIRAxEAPwDcaUpQKUpQKUpQK8t1ak80co/j1U/eU8j+B869VKCCOvmFgl2nZ55LKuTC3z6ofI/WptHBAIIIPMEcwa+Lq3SRSjqGVuRB6VRtQtLnTG7W3YyWpPvRtz2fp94fPzC/UqJ4f4ghu1yhw4+JD8Q/UedS1ApSlApSlApSlApSlApSlArqublIxud1RfFiAPqa4u7hY0eRvhRSx9AM1iuv6y9y5kkPovco8BWrV1ZzTPfEQwb29GtEdczLZbPU4JTiKaNyOoRlY/ga9dfmqa4aNg6MVdTkMpwQfI1t/s94iN7aCR8dqjdnJjvIAIb5gj55psa3i7ieYT1NvzR3HErNXDqCCCAQeRB6GuaVlbGb8QaC1nMJ7clUJ5EfuH+H0P8ASrfw5rguEwcCVfiHj5jy/KpO8tlkRo2GVYY/rVBjge3m5HDIfr/Qig0SldFjdCRA47+7wPeK76BSlKBSlKBSlKBSlKDxa3aGa3liHV42UepHL8a/O9xOykowIZSQQeoI5EV+laqvFPAVpet2jbo5T1ePA3feBGD69fOtetseLmJ+pYtvUjNxPuGA3dzW0+xfTHisWkcEdvJvUH+EAKD88E+mK+NI9klnE4eV5J8HIVsKnzA5n0zirdrOuW1mgMzhBj3UHNjjuVR/sK9z5vL/AFqlgwRhjmyUpWUar7XHzi3tRjuMpJJ/8V/WrLw3c6zMBJOtrCh5hGSQyY9A/u/Pn5VRbDasc26XVzVvPFe1yqv8Q2o3q3iMH5f71Ow7se9jP/LnH41EcSy4CfP+VVLXTw/Ltcxno3Meo/p+VWCqVb3eJEbwYfnzq60ClKUClKUClKUClKUClKieJNaW1hLnm55Ivif0HfUqUm9orH3KGTJXHWb2nqEfxlxUtomxMNOw5L3KP4m/kO+sZ1O4kmcySuXdupP/ANyHlUrfSPI7SOSzsckmo2ZK+9h1a4a/n25bN/IW2L/j1C5eyvhVXY3kq5VDiIHoWHV/l0HnnwrV68Gg2Qht4ogPgQA+uMsfmc176+JmyfO8y6bXx+PHEfsqk8WagDNsB5IMfM8z/KrFxFrC20Jc43Hki+J/QdTWVNeliWJySck+JNVLkx9o5j1rUayDSCZJ4kH7zgfLPP8ADNa/QKUpQKUpQKUpQKUpQcE1VNR4Ze8czTSmMdI4wM7V8W8z1I9BmrZSrMeW2Oea/anNgpmj437j/GHatbNDK8T/ABIcevgfmMVEyPzzVr9p1uY7zfjlIgYHzA2n8h9apUstdBTL88cWn3Dk763iy2rHqX6Nt5QyKw6MAR6EZFRmu8RQWq5dsv3RrzY/oPM1QOG9U+2Wy2n2gw3UQxGdxUSp3KcdSOnyB8ahdQ0W5iYiSGTPjgsD8xkGufyUmlprLrsWSMlItBrmsy3UpkkPkqjoo8B+teAOa9dvpNw5wkEh9Eb9Kl9P4eVZUjuGzK3w20ZDSt5uRyjQd7E+nOocJpn2a6OzMbpx7q5WPzJ5MfkOXzPhWiV128KoqoqhVUYAHQeQrsrx6UpSgUpSgUpSgUpSgUpSggOM+HBewbAQsqe9Gx6Z7wfI/pWGazp81s5jmjZG8+h8wehHmK/SVee+sYpl2SxpIvg4BH41pw7NscfH0yZ9SuWfl9S/L8k3eDzqSh461GJdqXkmB/EQ/wCLAmtku/ZjprnPZOme5JGA+hJr7sPZrpkR3fZ95/6jM4+hO38Ktvs0t6Qxa1qe2U6LNreqNsS4n7POGkLFIh6lcZ9Bk1sfBvCEGnxlU9+V/wC8lb4mPh5L5fXNT8MKooVFCqOQCgAD0Ar7rLfJ8uo6hrrXgpSlVplKUoFKUoFVbRuLmuL2a0FswWJnUy7srlG2jI28s+tWms14d1+4l1maBnxDG0wWNQFXk2Nxx8Td+TQd997UkjmkhFpIxjdkyGHPaSM4x5V7uG/aPb3UogaN4ZGOF3YKk+GR0PqKqHBX7dm+/P8Am1c+16ARX0EyDazKGJHLLI/JvXp9KDY6VnnH/FVylxDY2ZCyy7cvgEgucKoyCB4k+lR8Wu6hp9/FbXc4nil288dAzbcg4BBB7umKDU6VlWt8UaiNVe0t5AQWCIjKu0Fowck4zgEluvdXhXiLVrS++yySrPI+FCnGwlx7jA4BGCR4d9BsdKyJdf1Oy1GKC6nEokZdyjBQq7bcj3QQQc/SvVxZxVfxap9mt5AVJRVjKrtLOg6nGfiOetBqdVPUuM+yv0sfs5bcyDtN2AN/fjb3etVCLiDUrC/jgvJxKkhXOMFdrnbuU7QQQc8vKpPXeJbqPWYrVJsQM8QKbU5hsbuZGefrQaRSs04l4kvJ9QGnWUgi2nDPgZJC7mOcHAA5YHfXPC/EN5BqJ068lEueSvjmDt3rg4GQR3HvoNKpWNji7VJLy4toGMjF3SMbUGwB/izgdFGMscc6keIOJr+1gtbPd/bZQTI52sRukKoB+7k+Pdig1OlZRc61qWmXUKXdwJ4pcbuXQZw2DgEEZz4GtXoFZFwf+37n78//ALmtdqs6XwZDDeSXqySGSQuSp27RvOTjln8aCgcGMBrs2Tj35+vq1c+1W4W4v4IIiHZVCHac+879OXfjH1qQ1Hg/T5buUPNcqzO5L4j7LcFEzqDjPJWzzGOvOpXhrh/SbQrdLcCQ7tiPIykBsc8AADOOeT0HOgq3Htpu1mNGdo1k7IB1OCoPu5B7uYNe7VuHLGC7himvLqSclSgID9XwoJPTmKtnE+hWWpSdkZsXMQPNCNwGeYYHqASPMZ86heGeG7SG8kkklnmkt85lkAESlVUnJ3FshWGC2ByOM45BESf4k/zB/oV9cS/4hh+/D+QqzDR7BrxdTF2dzEuBldh2gRN3Z5ErnzIpqulWD3v297va8MiKy7l2B16KeWc8jnn3Ggq/tH/bNr6Q/wCq1fHEsqrxDGzHAEkOSfuqKuOscOWd5fLKblu3hCHs0K4wrbwTkZ/eHf0I8apPF9kk2uiGQEpI0atg4ODGooO/2ssH1G2jU5YIgIHcWkOBX1xN/iGH/uQ/yqR0XhK0tr87nnmkhAk5hAi+6WUk7tzH3TjA64rtuUsp7lNS/tPaLJEBEBHzJTfH34wVGT72aCsyaZ22uTQNK8W+R8Ohw3wlgAfMVN6ZollFqkcX2q5lukbPvAMvJN2C3XktSXGGh2dykWoCSWN5NgRowMsWwEyGIAI8dw/KvRwfw9Z2knbGZ2uZFH9/7rrvDN8PUEhWyST0NBWfZ9+2rn/O/wBUV1+1aDOpwbmKKyRjeORUdowJB8R1qz6FpthBeC6iuXke6aRFHulMn/iN0AI5Yx6ivVxda6bfQq8tyqhWKpKrDkcbivPkRjB9OYNBVOJuGLK3lijub27d35oMByPex39Mn8q12qBwvwLa9ot19qe62HCkkbQUOBnmSdp6DOKv9ApSlBVtZ4dUfaLjtGxslkEeBgO8HZM2epG0dPEmvNacILLbxFpmLEiQnauCGhSIrt6fCq8/GlKCbstBWOczh2JJkODjH/FKE/Tsx9a8l5woksssjyse1jePAVQQJAAQWAywGMqD0yaUoOi44NV1UGd9252dgq+/2jo7DH7vONennXquOGQVwszI3bvOG2qcGUMrDB5HkxxnyrmlB22PDyR3L3IckuPhIXkSqKTnrjCLy5d/liKv+ChJqK3/AG5BVlbs9mfgUD4t3fjwpSgk5uHg1y9z2mC6BGUohOArL7rEbl+LnjrXkt+DIFUR5zGHicoVXaxij7PmMY97kT50pQS2raYJoeyDBBy5bVZSB+6VIwVx3VXLXRkW+hgDMfs9mQGOCW3MUGRjHugvjr8X1UoJCx4UWNYl7Zj2cjOOQAw8XZFAP3Rjny6HPpXRHwTGIRD2rYBbmFUHDQG3GfEhTnPj5cqUoJzSNMWBXVWJ3yGTn3Fscvwr3UpQf//Z"
                     }
                  }
      responses:
        '200':
          description: OK. Return the updated record.
          content: {}
        '403':
          $ref: '#/components/responses/ForbiddenUpdateRecordRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - Models
      summary: delete record by id
      responses:
        '200':
          description: OK. Record successfully deleted.
          content:
            application/json:
               schema:
                  type: object
               example:
                  {
                     "msg": "Deleted"
                  }
        '403':
          $ref: '#/components/responses/ForbiddenDeleteRecordRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /models/{tableName}/{id}/{columnName}:
    parameters:
      - $ref: '#/components/parameters/locale'
    get:
      tags:
        - Models
      summary: Get property value of a record
      parameters:
        - name: tableName
          in: path
          schema:
            type: string
          description: table name
          required: true
          examples:
            AD_Client:
              value: 'ad_client'
        - name: id
          in: path
          schema:
            oneOf:
              - type: string
              - type: integer
          description: integer record id or string record uuid
          required: true
          examples:
            AD_Client_11:
              value: 11
        - name: columnName
          in: path
          schema:
            type: string
          description: column name
          required: true
          examples:
            AD_Client_AD_Language:
              value: AD_Language
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
        '403':
          $ref: '#/components/responses/ForbiddenGetRecordRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /models/{tableName}/{id}/attachments/zip:
    parameters:
      - $ref: '#/components/parameters/locale'
    get:
      tags:
        - Models
      summary: Get attachment of a record as zip archive
      parameters:
        - name: tableName
          in: path
          schema:
            type: string
          description: table name
          required: true
          examples:
            C_Order:
              value: 'c_order'
        - name: id
          in: path
          schema:
            oneOf:
              - type: string
              - type: integer
          description: integer record id or string record uuid
          required: true
          examples:
            C_Order_1000000:
              value: 1000000
      responses:
        '200':
          $ref: '#/components/responses/BinaryResponse'
        '403':
          $ref: '#/components/responses/ForbiddenGetRecordRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /models/{tableName}/{id}/attachments/{fileName}:
    parameters:
      - $ref: '#/components/parameters/locale'
    get:
      tags:
        - Models
      summary: Get attachment of a record by file name
      parameters:
        - name: tableName
          in: path
          schema:
            type: string
          description: table name
          required: true
        - name: id
          in: path
          schema:
            oneOf:
              - type: string
              - type: integer
          description: integer record id or string record uuid
          required: true
        - name: fileName
          in: path
          schema:
            type: string
          description: file name
          required: true
      responses:
        '200':
          $ref: '#/components/responses/BinaryResponse'
        '403':
          $ref: '#/components/responses/ForbiddenGetRecordRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...